pyproject.toml
[tool.poetry]
name = "aegis-shared"
version = "0.1.0"
description = "Shared library for the Aegis project - 이지스 시스템의 공통 기능을 제공하는 중앙 집중식 라이브러리"
authors = ["Aegis Team <team@aegis.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/aegis-project/aegis-shared"
repository = "https://github.com/aegis-project/aegis-shared"
documentation = "https://aegis-shared.readthedocs.io"
keywords = ["microservices", "shared-library", "fastapi", "sqlalchemy", "kafka", "redis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Database",
    "Topic :: System :: Distributed Computing",
]
packages = [{include = "aegis_shared"}]

[tool.poetry.dependencies]
python = "^3.11"
PyJWT = "^2.8.0"
fastapi = "^0.104.1"
structlog = "^23.1.0"
elasticsearch = "^8.11.0"
aiokafka = "^0.10.0"
prometheus-client = "^0.19.0"
opentelemetry-api = "^1.21.0"
opentelemetry-sdk = "^1.21.0"
opentelemetry-exporter-jaeger = "^1.21.0"
sentry-sdk = "^1.39.1"
redis = "^5.0.1"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"
pytz = "^2023.3.post1"
bcrypt = "^4.0.1"
cryptography = "^41.0.5"
python-magic = "^0.4.27"
PyYAML = "^6.0.1"
sqlalchemy = "^2.0.23"
tenacity = "^8.2.3"
python-json-logger = "^2.0.7"
uvicorn = "^0.24.0"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
aiosqlite = "^0.19.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

[tool.pytest.ini_options]
addopts = "--cov=aegis_shared --cov-report=term-missing --cov-report=html"
testpaths = ["aegis_shared/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["aegis_shared"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "aiokafka.*",
    "elasticsearch.*",
    "prometheus_client.*",
    "opentelemetry.*",
    "sentry_sdk.*",
    "magic.*",
]
ignore_missing_imports = true
